Rename android testbed_manual -> testbed_cocoapods
