+++ testbed/testbed_cocoapods/android/app/BUCK
@@ -46,13 +46,13 @@ android_library(
 
 android_build_config(
   name = 'build_config',
-  package = 'com.testbed_manual',
+  package = 'com.testbed_cocoapods',
 )
 
 android_resource(
   name = 'res',
   res = 'src/main/res',
-  package = 'com.testbed_manual',
+  package = 'com.testbed_cocoapods',
 )
 
 android_binary(
