+++ testbed/testbed_cocoapods/android/app/build.gradle
@@ -81,7 +81,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.testbed_manual"
+        applicationId "com.testbed_cocoapods"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
