+++ testbed/testbed_cocoapods/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">testbed_manual</string>
+    <string name="app_name">testbed_cocoapods</string>
 </resources>
